start() method taking all endpoints as an argument
