+++ cosmos-server/src/main/scala/com/mesosphere/cosmos/Cosmos.scala
@@ -3,33 +3,26 @@ package com.mesosphere.cosmos
 import _root_.io.circe.Json
 import _root_.io.finch._
 import _root_.io.finch.circe.dropNullKeys._
+import _root_.io.finch.internal.ToResponse
 import com.mesosphere.cosmos.app.Logging
 import com.mesosphere.cosmos.circe.Encoders._
 import com.mesosphere.cosmos.finch.EndpointHandler
-import com.mesosphere.cosmos.finch.FinchExtensions._
-import com.mesosphere.cosmos.finch.RequestValidators
 import com.mesosphere.cosmos.handler._
 import com.mesosphere.cosmos.repository.DefaultInstaller
 import com.mesosphere.cosmos.repository.DefaultUniverseInstaller
 import com.mesosphere.cosmos.repository.LocalPackageCollection
 import com.mesosphere.cosmos.repository.OperationProcessor
-import com.mesosphere.cosmos.repository.PackageSourcesStorage
 import com.mesosphere.cosmos.repository.SyncFutureLeader
 import com.mesosphere.cosmos.repository.Uninstaller
 import com.mesosphere.cosmos.repository.UniverseClient
 import com.mesosphere.cosmos.repository.ZkRepositoryList
 import com.mesosphere.cosmos.rpc.MediaTypes
-import com.mesosphere.cosmos.rpc.v1.circe.MediaTypedBodyParsers._
-import com.mesosphere.cosmos.rpc.v1.circe.MediaTypedEncoders._
-import com.mesosphere.cosmos.rpc.v1.circe.MediaTypedRequestDecoders._
-import com.mesosphere.cosmos.rpc.v2.circe.MediaTypedEncoders._
 import com.mesosphere.cosmos.storage.GarbageCollector
 import com.mesosphere.cosmos.storage.ObjectStorage
 import com.mesosphere.cosmos.storage.PackageStorage
 import com.mesosphere.cosmos.storage.StagedPackageStorage
 import com.mesosphere.cosmos.storage.installqueue.InstallQueue
 import com.mesosphere.cosmos.storage.installqueue.ReaderView
-import com.mesosphere.universe
 import com.netaporter.uri.Uri
 import com.twitter.app.App
 import com.twitter.app.Flag
@@ -53,114 +46,18 @@ import com.twitter.util.Try
 import org.slf4j.Logger
 import scala.concurrent.duration._
 
+final class Cosmos[A](allEndpoints: Endpoint[A])(implicit
+  statsReceiver: StatsReceiver = NullStatsReceiver,
+  tr: ToResponse.Aux[A, Application.Json],
+  tre: ToResponse.Aux[Exception, Application.Json]
+) {
 
-private[cosmos] final class Cosmos(
-  capabilitiesHandler: CapabilitiesHandler,
-  packageAddHandler: EndpointHandler[rpc.v1.model.AddRequest, rpc.v1.model.AddResponse],
-  packageDescribeHandler: EndpointHandler[rpc.v1.model.DescribeRequest, universe.v3.model.PackageDefinition],
-  packageInstallHandler: EndpointHandler[rpc.v1.model.InstallRequest, rpc.v2.model.InstallResponse],
-  packageListHandler: EndpointHandler[rpc.v1.model.ListRequest, rpc.v1.model.ListResponse],
-  packageListVersionsHandler: EndpointHandler[rpc.v1.model.ListVersionsRequest, rpc.v1.model.ListVersionsResponse],
-  packageRenderHandler: EndpointHandler[rpc.v1.model.RenderRequest, rpc.v1.model.RenderResponse],
-  packageRepositoryAddHandler: EndpointHandler[rpc.v1.model.PackageRepositoryAddRequest, rpc.v1.model.PackageRepositoryAddResponse],
-  packageRepositoryDeleteHandler: EndpointHandler[rpc.v1.model.PackageRepositoryDeleteRequest, rpc.v1.model.PackageRepositoryDeleteResponse],
-  packageRepositoryListHandler: EndpointHandler[rpc.v1.model.PackageRepositoryListRequest, rpc.v1.model.PackageRepositoryListResponse],
-  packageSearchHandler: EndpointHandler[rpc.v1.model.SearchRequest, rpc.v1.model.SearchResponse],
-  packageUninstallHandler: EndpointHandler[rpc.v1.model.UninstallRequest, rpc.v1.model.UninstallResponse],
-  serviceStartHandler: EndpointHandler[rpc.v1.model.ServiceStartRequest, rpc.v1.model.ServiceStartResponse],
-  extraEndpoints: Endpoint[Json]
-)(implicit statsReceiver: StatsReceiver = NullStatsReceiver) {
-
-  lazy val logger: Logger = org.slf4j.LoggerFactory.getLogger(classOf[Cosmos])
-
-  // Package Handlers
-  val packageInstall: Endpoint[Json] = {
-    route(post("package" :: "install"), packageInstallHandler)(RequestValidators.standard)
-  }
-
-  val packageUninstall: Endpoint[Json] = {
-    route(post("package" :: "uninstall"), packageUninstallHandler)(RequestValidators.standard)
-  }
-
-  val packageDescribe: Endpoint[Json] = {
-    route(post("package" :: "describe"), packageDescribeHandler)(RequestValidators.standard)
-  }
-
-  val packageRender: Endpoint[Json] = {
-    route(post("package" :: "render"), packageRenderHandler)(RequestValidators.standard)
-  }
-
-  val packageListVersions: Endpoint[Json] = {
-    route(
-      post("package" :: "list-versions"),
-      packageListVersionsHandler
-    )(RequestValidators.standard)
-  }
-
-  val packageSearch: Endpoint[Json] = {
-    route(post("package" :: "search"), packageSearchHandler)(RequestValidators.standard)
-  }
-
-  val packageList: Endpoint[Json] = {
-    route(post("package" :: "list"), packageListHandler)(RequestValidators.standard)
-  }
-
-  val packageRepositoryList: Endpoint[Json] = {
-    route(
-      post("package" :: "repository" :: "list"),
-      packageRepositoryListHandler
-    )(RequestValidators.standard)
-  }
-
-  val packageRepositoryAdd: Endpoint[Json] = {
-    route(
-      post("package" :: "repository" :: "add"),
-      packageRepositoryAddHandler
-    )(RequestValidators.standard)
-  }
-
-  val packageRepositoryDelete: Endpoint[Json] = {
-    route(
-      post("package" :: "repository" :: "delete"),
-      packageRepositoryDeleteHandler
-    )(RequestValidators.standard)
-  }
-
-  val packageAdd: Endpoint[Json] = {
-    route(post("package" :: "add"), packageAddHandler)(RequestValidators.selectedBody)
-  }
-
-  // Service Handlers
-  val serviceStart: Endpoint[Json] = {
-    route(post("service" :: "start"), serviceStartHandler)(RequestValidators.standard)
-  }
-
-  // Capabilities
-  val capabilities: Endpoint[Json] = {
-    route(get("capabilities"), capabilitiesHandler)(RequestValidators.noBody)
-  }
+  lazy val logger: Logger = org.slf4j.LoggerFactory.getLogger(classOf[Cosmos[_]])
 
   val service: Service[Request, Response] = {
     val stats = statsReceiver.scope("errorFilter")
 
-    val endpoints = (
-      // Keep alphabetized
-      capabilities
-        :+: packageAdd
-        :+: packageDescribe
-        :+: packageInstall
-        :+: packageList
-        :+: packageListVersions
-        :+: packageRender
-        :+: packageRepositoryAdd
-        :+: packageRepositoryDelete
-        :+: packageRepositoryList
-        :+: packageSearch
-        :+: packageUninstall
-        :+: serviceStart
-      ) :+: extraEndpoints
-
-    val api = endpoints.handle {
+    val api = allEndpoints.handle {
       case re: CosmosError =>
         stats.counter(s"definedError/${sanitiseClassName(re.getClass)}").incr()
         val output = Output.failure(
@@ -213,58 +110,6 @@ private[cosmos] final class Cosmos(
 
 }
 
-object Cosmos {
-
-  private[cosmos] def apply(
-    adminRouter: AdminRouter,
-    packageRunner: PackageRunner,
-    sourcesStorage: PackageSourcesStorage,
-    universeClient: UniverseClient,
-    installQueue: InstallQueue,
-    objectStorages: Option[(LocalPackageCollection, StagedPackageStorage)],
-    extraEndpoints: Endpoint[Json]
-  )(implicit statsReceiver: StatsReceiver = NullStatsReceiver): Cosmos = {
-
-    val repositories = new MultiRepository(
-      sourcesStorage,
-      universeClient
-    )
-
-    val packageAddHandler = enableIfSome(objectStorages, "package add") {
-      case (_, stagedStorage) => new PackageAddHandler(repositories, stagedStorage, installQueue)
-    }
-
-    val serviceStartHandler = enableIfSome(objectStorages, "service start") {
-      case (localPackageCollection, _) =>
-        new ServiceStartHandler(localPackageCollection, packageRunner)
-    }
-
-    new Cosmos(
-      new CapabilitiesHandler,
-      packageAddHandler,
-      new PackageDescribeHandler(repositories),
-      new PackageInstallHandler(repositories, packageRunner),
-      new ListHandler(adminRouter, uri => repositories.getRepository(uri)),
-      new ListVersionsHandler(repositories),
-      new PackageRenderHandler(repositories),
-      new PackageRepositoryAddHandler(sourcesStorage),
-      new PackageRepositoryDeleteHandler(sourcesStorage),
-      new PackageRepositoryListHandler(sourcesStorage),
-      new PackageSearchHandler(repositories),
-      new UninstallHandler(adminRouter, repositories),
-      serviceStartHandler,
-      extraEndpoints
-    )(statsReceiver)
-  }
-
-  private[this] def enableIfSome[A, Req, Res](requirement: Option[A], operationName: String)(
-    f: A => EndpointHandler[Req, Res]
-  ): EndpointHandler[Req, Res] = {
-    requirement.fold[EndpointHandler[Req, Res]](new NotConfiguredHandler(operationName))(f)
-  }
-
-}
-
 trait CosmosApp
 extends App
 with AdminHttpServer
@@ -274,12 +119,18 @@ with Lifecycle.Warmup
 with Stats
 with Logging {
 
-  protected val extraEndpoints: Endpoint[Json]
+  def main(): Unit
 
   lazy val logger: Logger = org.slf4j.LoggerFactory.getLogger(getClass)
 
-  final def main(): Unit = {
-    val service = startCosmos()(statsReceiver)
+  final def start[A](allEndpoints: Endpoint[A])(implicit
+    tr: ToResponse.Aux[A, Application.Json],
+    tre: ToResponse.Aux[Exception, Application.Json]
+  ): Unit = {
+    HttpProxySupport.configureProxySupport()
+    implicit val sr = statsReceiver
+
+    val service = LoggingFilter.andThen(new Cosmos(allEndpoints).service)
     val maybeHttpServer = startServer(service)
     val maybeHttpsServer = startTlsServer(service)
 
@@ -331,10 +182,15 @@ with Logging {
     }
   }
 
-  private[this] def startCosmos()(implicit stats: StatsReceiver): Service[Request, Response] = {
-    HttpProxySupport.configureProxySupport()
+}
+
+object Main extends CosmosApp {
+
+  override def main(): Unit = {
+    implicit val sr = statsReceiver
 
     val adminRouter = configureDcosClients().get
+    val packageRunner = new MarathonPackageRunner(adminRouter)
 
     val zkUri = zookeeperUri()
     logger.info("Using {} for the ZooKeeper connection", zkUri)
@@ -342,15 +198,15 @@ with Logging {
     val zkClient = zookeeper.Clients.createAndInitialize(zkUri)
     onExit(zkClient.close())
 
-    val repoList = ZkRepositoryList(zkClient)
-    onExit(repoList.close())
+    val sourcesStorage = ZkRepositoryList(zkClient)
+    onExit(sourcesStorage.close())
 
     val installQueue = InstallQueue(zkClient)
     onExit(installQueue.close())
 
-    val objectStorages = configureObjectStorage(installQueue)
+    val allObjectStorages = configureObjectStorage(installQueue)
 
-    for ((packageStorage, stageStorage, _) <- objectStorages) {
+    for ((packageStorage, stageStorage, _) <- allObjectStorages) {
       val processingLeader = SyncFutureLeader(
         zkClient,
         OperationProcessor(
@@ -368,23 +224,107 @@ with Logging {
       onExit(processingLeader.close())
     }
 
-    LoggingFilter.andThen(
-      Cosmos(
-        adminRouter,
-        new MarathonPackageRunner(adminRouter),
-        repoList,
-        UniverseClient(adminRouter),
-        installQueue,
-        objectStorages.map {
-          case (_, stagedStorage, localCollection) =>
-            (localCollection, stagedStorage)
-        },
-        extraEndpoints
-      ).service
+    val objectStorages = allObjectStorages.map {
+      case (_, stagedStorage, localCollection) =>
+        (localCollection, stagedStorage)
+    }
+
+    val universeClient = UniverseClient(adminRouter)
+
+    val repositories = new MultiRepository(
+      sourcesStorage,
+      universeClient
     )
+
+    val packageAddHandler = enableIfSome(objectStorages, "package add") {
+      case (_, stagedStorage) => new PackageAddHandler(repositories, stagedStorage, installQueue)
+    }
+
+    val serviceStartHandler = enableIfSome(objectStorages, "service start") {
+      case (localPackageCollection, _) =>
+        new ServiceStartHandler(localPackageCollection, packageRunner)
+    }
+
+    val cosmosApi = new CosmosApi(
+      new CapabilitiesHandler,
+      packageAddHandler,
+      new PackageDescribeHandler(repositories),
+      new PackageInstallHandler(repositories, packageRunner),
+      new ListHandler(adminRouter, uri => repositories.getRepository(uri)),
+      new ListVersionsHandler(repositories),
+      new PackageRenderHandler(repositories),
+      new PackageRepositoryAddHandler(sourcesStorage),
+      new PackageRepositoryDeleteHandler(sourcesStorage),
+      new PackageRepositoryListHandler(sourcesStorage),
+      new PackageSearchHandler(repositories),
+      new UninstallHandler(adminRouter, repositories),
+      serviceStartHandler
+    )
+
+    import cosmosApi._
+
+    // Keep alphabetized
+    val allEndpoints = (
+      capabilities
+        :+: packageAdd
+        :+: packageDescribe
+        :+: packageInstall
+        :+: packageList
+        :+: packageListVersions
+        :+: packageRender
+        :+: packageRepositoryAdd
+        :+: packageRepositoryDelete
+        :+: packageRepositoryList
+        :+: packageSearch
+        :+: packageUninstall
+        :+: serviceStart
+      )
+
+    start(allEndpoints)
   }
 
-  private[this] def configureDcosClients(): Try[AdminRouter] = {
+  private def enableIfSome[A, Req, Res](requirement: Option[A], operationName: String)(
+    f: A => EndpointHandler[Req, Res]
+  ): EndpointHandler[Req, Res] = {
+    requirement.fold[EndpointHandler[Req, Res]](new NotConfiguredHandler(operationName))(f)
+  }
+
+  private def configureObjectStorage(
+    installQueue: ReaderView
+  )(
+    implicit statsReceiver: StatsReceiver
+  ): Option[(PackageStorage, StagedPackageStorage, LocalPackageCollection)] = {
+    def fromFlag(
+      flag: Flag[ObjectStorageUri],
+      description: String
+    ): Option[ObjectStorage] = {
+      val value = flag.get.map(_.toString).getOrElse("None")
+      logger.info(s"Using {} for the $description storage URI", value)
+      flag.get.map(uri => ObjectStorage.fromUri(uri)(statsReceiver))
+    }
+
+    (
+      fromFlag(packageStorageUri, "package").map(PackageStorage(_)),
+      fromFlag(stagedPackageStorageUri, "staged package").map(StagedPackageStorage(_))
+    ) match {
+      case (Some(packageStorage), Some(stagedStorage)) =>
+        Some((packageStorage, stagedStorage, LocalPackageCollection(packageStorage, installQueue)))
+      case (None, None) =>
+        None
+      case (Some(_), None) =>
+        throw new IllegalArgumentException(
+          "Missing staged storage configuration. Staged storage configuration required if " +
+            "package storage provided."
+        )
+      case (None, Some(_)) =>
+        throw new IllegalArgumentException(
+          "Missing package storage configuration. Package storage configuration required if " +
+            "stage storage provided."
+        )
+    }
+  }
+
+  private def configureDcosClients(): Try[AdminRouter] = {
     import com.netaporter.uri.dsl._
 
     Try(dcosUri())
@@ -431,43 +371,4 @@ with Logging {
       }
   }
 
-  private[this] def configureObjectStorage(
-    installQueue: ReaderView
-  )(
-    implicit statsReceiver: StatsReceiver
-  ): Option[(PackageStorage, StagedPackageStorage, LocalPackageCollection)] = {
-    def fromFlag(
-      flag: Flag[ObjectStorageUri],
-      description: String
-    ): Option[ObjectStorage] = {
-      val value = flag.get.map(_.toString).getOrElse("None")
-      logger.info(s"Using {} for the $description storage URI", value)
-      flag.get.map(uri => ObjectStorage.fromUri(uri)(statsReceiver))
-    }
-
-    (
-      fromFlag(packageStorageUri, "package").map(PackageStorage(_)),
-      fromFlag(stagedPackageStorageUri, "staged package").map(StagedPackageStorage(_))
-    ) match {
-      case (Some(packageStorage), Some(stagedStorage)) =>
-        Some((packageStorage, stagedStorage, LocalPackageCollection(packageStorage, installQueue)))
-      case (None, None) =>
-        None
-      case (Some(_), None) =>
-        throw new IllegalArgumentException(
-          "Missing staged storage configuration. Staged storage configuration required if " +
-          "package storage provided."
-        )
-      case (None, Some(_)) =>
-        throw new IllegalArgumentException(
-          "Missing package storage configuration. Package storage configuration required if " +
-          "stage storage provided."
-        )
-    }
-  }
-
-}
-
-object Main extends CosmosApp {
-  override val extraEndpoints: Endpoint[Json] = Endpoint.empty
 }
